refactor : hotPlace select test 완료

+++ backend/src/main/java/com/ssafy/enjoytrip/core/hotplace/model/dto/HotPlaceArticleResponse.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class HotPlaceArticleResponse {
 
-    private int hotPlaceArticleId;
+    private Long hotPlaceArticleId;
     private String hotPlaceId;
     private String hotPlaceName;
     private String userId;
